backend/requirements.txt
annotated-types==0.5.0
asgiref==3.7.2
Django==4.2.5
django-environ==0.11.2
djangorestframework==3.14.0
Faker==19.6.2
faker-marketdata==0.4
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
psycopg2-binary==2.9.8
pycountry==22.3.5
pydantic==2.4.2
pydantic_core==2.10.1
pytest==7.4.2
pytest-asyncio==0.21.1
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
